move ConcurrentTaskRunnerG to estimate. minor cleanup
