+++ core/src/testFixtures/kotlin/org/cryptobiotic/rlauxe/workflow/WorkflowTasks.kt
@@ -1,7 +1,7 @@
 package org.cryptobiotic.rlauxe.workflow
 
-import org.cryptobiotic.rlauxe.concur.ConcurrentTaskG
-import org.cryptobiotic.rlauxe.concur.ConcurrentTaskRunnerG
+import org.cryptobiotic.rlauxe.estimate.ConcurrentTaskG
+import org.cryptobiotic.rlauxe.estimate.ConcurrentTaskRunnerG
 import org.cryptobiotic.rlauxe.core.Cvr
 import org.cryptobiotic.rlauxe.core.TestH0Status
 import org.cryptobiotic.rlauxe.oneaudit.makeContestOA
