+++ plots/src/test/kotlin/org/cryptobiotic/rlauxe/concur/RepeatedWorkflowRunner.kt
@@ -1,6 +1,7 @@
 
 package org.cryptobiotic.rlauxe.concur
 
+import org.cryptobiotic.rlauxe.estimate.ConcurrentTaskG
 import org.cryptobiotic.rlauxe.workflow.WorkflowResult
 import org.cryptobiotic.rlauxe.workflow.WorkflowTaskGenerator
