+++ plots/src/test/kotlin/org/cryptobiotic/rlauxe/raire/GenRaireNoErrorsPlots.kt
@@ -1,6 +1,6 @@
 package org.cryptobiotic.rlauxe.raire
 
-import org.cryptobiotic.rlauxe.concur.ConcurrentTaskG
+import org.cryptobiotic.rlauxe.estimate.ConcurrentTaskG
 import org.cryptobiotic.rlauxe.concur.RepeatedWorkflowRunner
 import org.cryptobiotic.rlauxe.rlaplots.showSampleSizesVsMargin
 import org.cryptobiotic.rlauxe.rlaplots.ScaleType
