+++ plots/src/test/kotlin/org/cryptobiotic/rlauxe/unittest/TestGenWorkflowTasks.kt
@@ -1,6 +1,6 @@
 package org.cryptobiotic.rlauxe.unittest
 
-import org.cryptobiotic.rlauxe.concur.ConcurrentTaskG
+import org.cryptobiotic.rlauxe.estimate.ConcurrentTaskG
 import org.cryptobiotic.rlauxe.concur.RepeatedWorkflowRunner
 import org.cryptobiotic.rlauxe.rlaplots.WorkflowResultsIO
 import org.cryptobiotic.rlauxe.rlaplots.WorkflowResultsPlotter
